+++ src/main/java/net/meander/subtlyd/mixin/client/blocks/BlockMixin.java
@@ -8,9 +8,7 @@ import net.meander.subtlyd.world.level.block.sounds.AmbientBushBlockSoundsPlayer
 import net.minecraft.core.BlockPos;
 import net.minecraft.util.RandomSource;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.level.block.AirBlock;
 import net.minecraft.world.level.block.Block;
-import net.minecraft.world.level.block.BushBlock;
 import net.minecraft.world.level.block.state.BlockState;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
@@ -22,9 +20,7 @@ import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 public class BlockMixin {
     @Inject(method = "animateTick", at = @At("HEAD"))
     private void animateTick(BlockState state, Level level, BlockPos pos, RandomSource random, CallbackInfo ci) {
-        if (state.is(BlockTagsSD.HAS_AMBIENT_BLOCK_SOUNDS)) {
-            playAmbientSounds(state.getBlock(), level, pos, random);
-        }
+        playAmbientSounds(state.getBlock(), level, pos, random);
     }
 
     /**
@@ -35,9 +31,11 @@ public class BlockMixin {
      * @param randomSource A randomSource type to determine the likelihood of sounds playing.
      */
     private void playAmbientSounds(Block block, Level level, BlockPos blockPos, RandomSource randomSource) {
-        if (block instanceof AirBlock) {
+        if (block.defaultBlockState().is(BlockTagsSD.TRIGGERS_AMBIENT_WIND_BLOCK_SOUNDS)) {
             AmbientAirBlockSoundsPlayer.playColdWindSounds(level, blockPos, randomSource);
-        } else if (block instanceof BushBlock) {
+        }
+
+        if (block.defaultBlockState().is(BlockTagsSD.TRIGGERS_AMBIENT_BUSH_BLOCK_SOUNDS)) {
             AmbientBushBlockSoundsPlayer.playAmbientBushSounds(level, blockPos, randomSource);
         }
     }